Rewrite trail vertex list building.
Most important change is to use the arrays directly rather than using
element indices. I don't remember why I did it that way as it gained
nothing and using arrays directly is much simpler.

Also, the "dummy" vertices at either end of the trail are no longer
duplicates of the end vertices, but are now extrapolations. This fixes up
the weirdness at either end of the trails.